Refactoring Control.cs , Logger.cs and cleaning other classes.
