+++ src/Machina/Interfaces.cs
@@ -1,16 +1,9 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
+﻿namespace Machina;
 
-namespace Machina
+/// <summary>
+/// Makes an object serializable into the Machina API instruction that would generate an equal instance of it.
+/// </summary>
+interface IInstructable
 {
-    /// <summary>
-    /// Makes an object serializable into the Machina API instruction that would generate an equal instance of it.
-    /// </summary>
-    interface IInstructable
-    {
-        string ToInstruction();
-    }
+    string ToInstruction();
 }
